Move test to its respective package for parser
